Remove --all flag from bosh clean-up on delete-deployment task
This was causing issues with our acceptance pipeline where we would
delete all deployments, clean-up --all, and redeploy. On the redeploy,
the expected bosh-dns release in the runtime-config was unable to be
found.

We learned that currently runtime-config releases are only uploaded when
you run `update-runtime-config`, not at actual deploy time.

We decided to remove the --all flag in order to work around this issue.